% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{check_stability}
\alias{check_stability}
\title{Vérifie la condition de stabilité du schéma explicite
D * dt / dx^2 < 1/2}
\usage{
check_stability(D, dx, dt)
}
\value{
list(ok = TRUE/FALSE, r = value, dt_max = valeur recommandée)
}
\description{
Vérifie la condition de stabilité du schéma explicite
D * dt / dx^2 < 1/2
}
