% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{concentration_profile}
\alias{concentration_profile}
\title{Génère un profil de concentration linéaire entre C1 et C2 dans un tissu}
\usage{
concentration_profile(C1, C2, x)
}
\arguments{
\item{C1}{numeric Concentration au point x_min}

\item{C2}{numeric Concentration au point x_max}

\item{x}{numeric Vecteur des positions (ordre quelconque)}
}
\value{
numeric Vecteur des concentrations C(x)
}
\description{
Génère un profil de concentration linéaire entre C1 et C2 dans un tissu
}
