% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{diffusion_coefficient_estimate}
\alias{diffusion_coefficient_estimate}
\title{Estimation du coefficient D à partir d'un tableau d'évolution (matrice temps x espace)}
\usage{
diffusion_coefficient_estimate(evolution, x, times)
}
\arguments{
\item{evolution}{matrix (nt x nx)}

\item{x}{numeric positions}

\item{times}{numeric vecteur temps (longueur = nrow(evolution))}
}
\value{
list(D_est = ..., lm = object, var_vs_time = data.frame)
}
\description{
Estimation du coefficient D à partir d'un tableau d'évolution (matrice temps x espace)
}
