% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{ficks_law_flux}
\alias{ficks_law_flux}
\title{Calcul du flux diffusif selon la 1ère loi de Fick (biomedicale)}
\usage{
ficks_law_flux(D, dC, dx = 1, S = 1)
}
\arguments{
\item{D}{numeric Coefficient de diffusion (cm^2/s ou m^2/s)}

\item{dC}{numeric Différence de concentration (C2 - C1)}

\item{dx}{numeric Distance (épaisseur de la membrane en cm)}

\item{S}{numeric surface de diffusion (cm^2)}
}
\value{
numeric Flux diffusif J (mg/s)
}
\description{
J = - D * S * (dC / dx)
}
