% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_concentration}
\alias{plot_concentration}
\title{Visualiser concentration (profil ou évolution)}
\usage{
plot_concentration(
  C,
  x = NULL,
  times = NULL,
  main = "Concentration",
  col = NULL
)
}
\arguments{
\item{C}{numeric vecteur (espace) ou matrice (temps x espace)}

\item{x}{numeric positions (si NULL, 1:ncol(C) pour matrice, 1:length(C) pour vecteur)}

\item{times}{numeric vecteur de temps (optionnel, longueur = nrow(C))}

\item{main}{title}

\item{col}{palette ou couleur}
}
\description{
Visualiser concentration (profil ou évolution)
}
