% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_diffusion}
\alias{simulate_diffusion}
\title{Lancer une simulation complète (profil initial + évolution + flux)}
\usage{
simulate_diffusion(
  C1 = 1,
  C2 = 0,
  dose = 1,
  x = seq(0, 1, length.out = 101),
  D = 0.01,
  dx = NULL,
  dt = 5e-04,
  steps = 200,
  bc = list(type = "dirichlet", left = C1, right = C2)
)
}
\arguments{
\item{C1}{numeric concentration gauche}

\item{C2}{numeric concentration droite}

\item{x}{numeric vecteur positions}

\item{D}{numeric coeff de diffusion}

\item{dx}{numeric pas spatial (si NULL calculé depuis x)}

\item{dt}{numeric pas temporel}

\item{steps}{integer nb d'itérations}

\item{bc}{list conditions aux limites (passé à fick_time_evolution)}
}
\value{
liste avec profils, flux, evolution (matrice) et le plot (NULL)
}
\description{
Lancer une simulation complète (profil initial + évolution + flux)
}
